autoconf: handle the soft-linked binaries
If the autoconf binaries are linked to somewhere else, the
solution with relative directories won't work well as the
relative directories aren't right in the case.

Extra handling with readlink can return the actual location
of the binaries and the referred directories can be found
correctly. readlink will be failed and return null when the
file isn't a link. In the case, the OR-ed file will be used.

Change-Id: Ib9121d98b647773da20c71d9d6f2acd0f8134015